+++ packages/solid/src/components/select/select-root.tsx
@@ -48,6 +48,7 @@ export const SelectRoot = <T extends CollectionItem>(props: SelectRootProps<T>) 
     'open',
     'positioning',
     'readOnly',
+    'required',
     'scrollToIndexFn',
     'value',
   ])