+++ packages/react/src/components/signature-pad/examples/with-field.tsx
@@ -1,16 +1,24 @@
+import { useState } from 'react'
 import { Field, SignaturePad } from '../..'
 
-export const WithField = (props: Field.RootProps) => (
-  <Field.Root {...props}>
-    <SignaturePad.Root>
-      <SignaturePad.Label>Label</SignaturePad.Label>
-      <SignaturePad.Control>
-        <SignaturePad.Segment />
-        <SignaturePad.ClearTrigger>Clear</SignaturePad.ClearTrigger>
-        <SignaturePad.Guide />
-      </SignaturePad.Control>
-    </SignaturePad.Root>
-    <Field.HelperText>Additional Info</Field.HelperText>
-    <Field.ErrorText>Error Info</Field.ErrorText>
-  </Field.Root>
-)
+export const WithField = (props: Field.RootProps) => {
+  const [value, setValue] = useState('')
+
+  return (
+    <Field.Root {...props}>
+      <SignaturePad.Root
+        onDrawEnd={(details) => details.getDataUrl('image/png').then((url) => setValue(url))}
+      >
+        <SignaturePad.Label>Label</SignaturePad.Label>
+        <SignaturePad.Control>
+          <SignaturePad.Segment />
+          <SignaturePad.ClearTrigger>Clear</SignaturePad.ClearTrigger>
+          <SignaturePad.Guide />
+        </SignaturePad.Control>
+        <SignaturePad.HiddenInput value={value} />
+      </SignaturePad.Root>
+      <Field.HelperText>Additional Info</Field.HelperText>
+      <Field.ErrorText>Error Info</Field.ErrorText>
+    </Field.Root>
+  )
+}