+++ packages/vue/src/components/signature-pad/examples/with-field.vue
@@ -1,16 +1,24 @@
 <script setup lang="ts">
-import { Field, SignaturePad } from '../..'
+import { ref } from 'vue'
+import { Field, SignaturePad, type SignaturePadDrawEndDetails } from '../..'
+
+const imageUrl = ref('')
+
+const handleDrawEnd = async (details: SignaturePadDrawEndDetails) => {
+  imageUrl.value = await details.getDataUrl('image/png')
+}
 </script>
 
 <template>
   <Field.Root>
-    <SignaturePad.Root>
+    <SignaturePad.Root  @draw-end="handleDrawEnd">
       <SignaturePad.Label>Label</SignaturePad.Label>
       <SignaturePad.Control>
         <SignaturePad.Segment />
         <SignaturePad.ClearTrigger>Clear</SignaturePad.ClearTrigger>
         <SignaturePad.Guide />
       </SignaturePad.Control>
+      <SignaturePad.HiddenInput :value="imageUrl" />
     </SignaturePad.Root>
     <Field.HelperText>Additional Info</Field.HelperText>
     <Field.ErrorText>Error Info</Field.ErrorText>